+++ src/realm/util/file.cpp
@@ -1288,10 +1288,15 @@ File::UniqueID File::get_unique_id() const
 #else // POSIX version
     struct stat statbuf;
     if (::fstat(m_fd, &statbuf) == 0) {
-        return {static_cast<uint_fast64_t>(statbuf.st_dev), static_cast<uint_fast64_t>(statbuf.st_ino)};
+        return {statbuf.st_dev, statbuf.st_ino};
     }
     throw std::system_error(errno, std::system_category(), "fstat() failed");
 #endif
+}
+
+FileDesc File::get_descriptor() const
+{
+    return m_fd;
 }
 
 bool File::get_unique_id(const std::string& path, File::UniqueID& uid)